Update some gems that were reported as having security vulnerabilities.
I tried to update bootstrap-sass because of
https://nvd.nist.gov/vuln/detail/CVE-2019-8331 but the latest version
changes from the sass gem to the sassc gem which causes Bootstrap to
fail to load with a sassc syntax error about not being able to find the
file. I spent enough time trying to figure out how to fix it and gave
up.